+++ src/components/code-bubble/code-bubble.ts
@@ -173,17 +173,36 @@ export default class CodeBubble extends LitElement {
     if (this.hasAttribute('framework')) {
       this.framework = this.getAttribute('framework') as string;
     }
+  }
 
-    // wait for code formatters to finish executing
-    setTimeout(() => {
-      this.getCode();
-      this.createPreview();
-    }, 10);
+  protected firstUpdated() {
+    this.loadCode();
+    this.initObserver();
+  }
+
+  private loadCode() {
+    this.getCode();
+    this.createPreview();
+  }
+
+  private initObserver() {
+    const observer = new MutationObserver(() => {
+      this.reloadCode();
+      const preview = this.querySelector('[slot="preview"]');
+      if (preview?.textContent) {
+        observer.disconnect();
+      }
+    });
+    observer.observe(this, { childList: true, subtree: true });
+  }
+
+  private reloadCode() {
+    this.getCode();
+    this.updatePreview();
   }
 
   private handleSlotChange() {
-    this.getCode();
-    this.updatePreview();
+    this.reloadCode();
   }
 
   private createPreview() {
@@ -279,8 +298,6 @@ export default class CodeBubble extends LitElement {
     element?: HTMLElement | null,
     language: string = 'html',
   ) {
-
-    console.log('LANG', language);
     if (!element) {
       return;
     }